+++ cmd/generate/prometheus/rule/rule.go
@@ -6,7 +6,7 @@ import (
 	"os"
 
 	"github.com/ajalab/slom/cmd/common"
-	configspec "github.com/ajalab/slom/internal/config/spec/native/v1alpha"
+	configspec "github.com/ajalab/slom/internal/config/spec/native/v1alpha1"
 	"github.com/ajalab/slom/internal/print"
 	"github.com/ajalab/slom/internal/prometheus/rule"
 	"github.com/ajalab/slom/internal/spec"
