+++ internal/config/spec/core/v1alpha1/config.go
@@ -1,4 +1,4 @@
-package v1alpha
+package v1alpha1
 
 // SpecConfig is a configuration for an SLO specification.
 // An SLO specification typically corresponds to a specific service or user journey and may include multiple SLO entries.
