+++ internal/config/spec/native/v1alpha1/parse.go
@@ -1,4 +1,4 @@
-package v1alpha
+package v1alpha1
 
 import (
 	"io"
